requirements.txt
mcp[cli]>=1.0.0
fastapi>=0.103.0
uvicorn>=0.23.2
httpx>=0.24.1
python-dotenv>=1.0.0
